added user management

+++ backend/src/main/java/de/windalert/service/CrudService.java
@@ -1,6 +1,7 @@
 package de.windalert.service;
 
 import de.windalert.domain.Spot;
+import de.windalert.domain.User;
 import de.windalert.domain.WindWindow;
 import de.windalert.dto.CreateSpotDTO;
 import de.windalert.dto.InfoDTO;
@@ -29,28 +30,28 @@ public class CrudService {
     }
 
 
-    public List<InfoDTO> getInfoList() {
-        return spotRepository.findAll().stream().sorted(Comparator.comparing(Spot::getId))
+    public List<InfoDTO> getInfoList(User user) {
+        return spotRepository.findAllByUser(user).stream().sorted(Comparator.comparing(Spot::getId))
                 .map(spot -> new InfoDTO(spot.getId(), spot.getName(), spot.getLatitude(),
                         spot.getLongitude(),
                         spot.getWindWindows().stream().sorted(Comparator.comparing(WindWindow::getId)).map((ww) -> new InfoDTO.WindWindow(ww.getId(),
                                 ww.getSpeed(), ww.getStartAngle(), ww.getEndAngle())).toList())).toList();
     }
 
-    public SpotDTO createSpot(CreateSpotDTO createSpotDTO) {
-        Optional<Spot> optionalSpot = spotRepository.findByName(createSpotDTO.name());
+    public SpotDTO createSpot(CreateSpotDTO createSpotDTO, User user) {
+        Optional<Spot> optionalSpot = spotRepository.findByNameAndUser(createSpotDTO.name(), user);
         if (optionalSpot.isPresent()) {
             throw new IllegalArgumentException("Spot does already exist");
         }
         log.info("Creating new spot with name {}", createSpotDTO.name());
-        Spot spot = new Spot(createSpotDTO.name(), new HashSet<>(),
+        Spot spot = new Spot(user, createSpotDTO.name(), new HashSet<>(),
                 createSpotDTO.latitude(), createSpotDTO.longitude());
         spot = spotRepository.save(spot);
         return new SpotDTO(spot.getId(), spot.getName(), spot.getLatitude(), spot.getLongitude());
     }
 
-    public SpotDTO updateSpot(SpotDTO spotDTO) {
-        Optional<Spot> optionalSpot = spotRepository.findById(spotDTO.id());
+    public SpotDTO updateSpot(SpotDTO spotDTO, User user) {
+        Optional<Spot> optionalSpot = spotRepository.findByIdAndUser(spotDTO.id(), user);
         if (optionalSpot.isEmpty()) {
             throw new IllegalArgumentException("Spot does not exist");
         }
@@ -65,19 +66,24 @@ public class CrudService {
         return new SpotDTO(spot.getId(), spot.getName(), spot.getLatitude(), spot.getLongitude());
     }
 
-    public void deleteSpot(Long id) {
+    public void deleteSpot(Long id, User user) {
+        Optional<Spot> optionalSpot = spotRepository.findByIdAndUser(id, user);
+        if (optionalSpot.isEmpty()) {
+            throw new IllegalArgumentException("Spot does not exist");
+        }
         log.info("Deleting spot with id {}", id);
         spotRepository.deleteById(id);
     }
 
-    public WindWindowDTO createWindWindow(WindWindowDTO windWindowDTO) {
-        Optional<Spot> optionalSpot = spotRepository.findById(windWindowDTO.spotId());
+    public WindWindowDTO createWindWindow(WindWindowDTO windWindowDTO, User user) {
+        Optional<Spot> optionalSpot = spotRepository.findByIdAndUser(windWindowDTO.spotId(), user);
         if (optionalSpot.isEmpty()) {
             throw new IllegalArgumentException("Spot does not exist");
         }
         log.info("Creating new wind window for spot with id {}", windWindowDTO.spotId());
         Spot spot = optionalSpot.get();
-        WindWindow windWindow = new WindWindow(windWindowDTO.speed(), windWindowDTO.startAngle(), windWindowDTO.endAngle());
+        WindWindow windWindow = new WindWindow(windWindowDTO.speed(), windWindowDTO.startAngle(), windWindowDTO.endAngle(),
+                spot);
         spot.getWindWindows().add(windWindow);
         windWindow = windWindowRepository.save(windWindow);
         return new WindWindowDTO(windWindow.getId(), windWindow.getSpeed(), windWindow.getStartAngle(),
@@ -85,12 +91,12 @@ public class CrudService {
 
     }
 
-    public WindWindowDTO updateWindWindow(WindWindowDTO windWindowDTO) {
+    public WindWindowDTO updateWindWindow(WindWindowDTO windWindowDTO, User user) {
         Optional<WindWindow> optionalWindWindow = windWindowRepository.findById(windWindowDTO.id());
         if (optionalWindWindow.isEmpty()) {
             throw new IllegalArgumentException("Wind window does not exist");
         }
-        Optional<Spot> optionalSpot = spotRepository.findById(windWindowDTO.spotId());
+        Optional<Spot> optionalSpot = spotRepository.findByIdAndUser(windWindowDTO.spotId(), user);
         if (optionalSpot.isEmpty()) {
             throw new IllegalArgumentException("Spot does not exist");
         }
@@ -105,7 +111,16 @@ public class CrudService {
                 windWindow.getEndAngle(), spot.getId());
     }
 
-    public void deleteWindWindow(Long id) {
+    public void deleteWindWindow(Long id, User user) {
+        Optional<WindWindow> optionalWindWindow = windWindowRepository.findById(id);
+        if (optionalWindWindow.isEmpty()) {
+            throw new IllegalArgumentException("Wind window does not exist");
+        }
+        WindWindow windWindow = optionalWindWindow.get();
+        Optional<Spot> optionalSpot = spotRepository.findByIdAndUser(windWindow.getSpot().getId(), user);
+        if (optionalSpot.isEmpty()) {
+            throw new IllegalArgumentException("Spot does not exist");
+        }
         log.info("Deleting wind window with id {}", id);
         windWindowRepository.deleteById(id);
     }
